% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addFirstTwo.R
\name{addFirstTwo}
\alias{addFirstTwo}
\title{addFirstTwo}
\usage{
addFirstTwo(vec)
}
\arguments{
\item{vec}{- a vector of any length}

\item{vec[1]}{- the first elements of the vector}

\item{vec[2]}{- the second element of the vector}
}
\value{
Returns the sum of the two elements of a vector
}
\description{
Adds the first two elements of a vector of any length.
}
