% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateRFM.R
\name{calculateRFM}
\alias{calculateRFM}
\title{calculateRFM}
\usage{
calculateRFM(
  data,
  weight_recency = 1,
  weight_frequency = 1,
  weight_monetary = 1
)
}
\value{
Returns a data containing the recency, frequency, monetary scores and the weighted final score for each customer
}
\description{
Calculates the weighted RFM score : recency, frequency, monetary for every customer
}
\details{
\code{data} contains the transactions data. The data set must contain a
            column labeled "Customer" that allows unique customer identification
            and a column labeled "TransDate", indicating the purchase date.
            The column "PurchAmount" specifies the total spending per purchase.
}
