\name{myPackage-package}
\alias{myPackage-package}
\alias{myPackage}
\docType{package}
\title{
\packageTitle{myPackage}
}
\description{
\packageDescription{myPackage}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{myPackage}
\packageIndices{myPackage}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{myPackage}

Maintainer: \packageMaintainer{myPackage}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
# simple examples of the most important functions
}
